refactored to new component ReviewSeller from ViewSellerProfile
