+++ frontend/pfiniture/src/components/my-account/order-history/Order.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import ViewSellerProfile from '../../landing/ViewSellerProfile'
-import client from '../../../API/api';
 
 const Order = (props) => {
     let order = props.order
